5.1.7
carbon copy of 5.1.6 with no info except for .attr("type", "text")  basically sets the attribute to look like this  type="text"  when clicked. on.  Once the blur comes in it over rides it by putting this line
$(".list-group").on("blur", "input[type='text']", function() {
to target it it by the type we just gave it and then futher down it gets changed back to attr("id")