5.2.4
Bootstrap focusing on columns and rows  and the different media break points you can put in to adjust.  Ex

  <div class="m-5 row justify-content-around">
this is our ROW <div>

        <!-- todo start -->
        <div class="col-12 col-md-6 col-xl-3 mb-3">
In here is where we define the column size within that row.  Every row is considered a size of 12 so you have to make 12 fit.  So col-12 will display one item per row at <576px.  Col-md-6 says at the will display 2 items per row at > than or = to 768 etc......
          <h2>To Do</h2>
          <ul id="list-toDo" class="list-group"></ul>
        </div>