5.2.6
Bootstrapped in colors.  Learned how to remove an element d-none and  then use d-lg-block to allow it to show up at that size.

Created a new div for the trash in the <main> and added the d-flex and flex column to set up the push for the trash.  Then added mt-auto to the trash so it will push itself to the bottom of the <main> ie screen.

Sticky top was a quick way to make an <el> stay on the top while scrolling.