5.3.5
A ton of info using .sortable.  Let's try and explain.

$(".card .list-group").sortable({
grabs the entire <div> for .card  (ex To Do) and zeroes in on the <ul> by the class .list-group.  All 4 share these 2 classes.  Then assigns the .sortable jquery.

  connectWith: $(".card .list-group"),
then defines how they are connected

  scroll: false,
  tolerance: "pointer",
  helper: "clone",
assigns some things that I will need to understand later

  activate: function (event) {
    console.log("activate", this);
  },
when clicked this logs

  deactivate: function (event) {
    console.log("deactivate", this);
  },
when the click is removed this logs

  over: function (event) {
    console.log("over", event.target);
  },
when dragged over with clicked item this logs

  out: function (event) {
    console.log("out", event.target);
  },
when dropped this logs

  update: function (event) {
    // array to store the task data in
    var tempArr = [];

    // loop over current set of children in sortable list
    $(this).children().each(function () {
      var text = $(this)
        .find("p")
        .text()
        .trim();

      var date = $(this)
        .find("span")
        .text()
        .trim();

      // add task data to the temp array as an object
      tempArr.push({
        text: text,
        date: date
      });
    });

    // trim down list's ID to match object property
    var arrName = $(this)
      .attr("id")
      .replace("list-", "");

    // update array on tasks object and save
    tasks[arrName] = tempArr;
    saveTasks();
  }
});

var createTask = function (taskText, taskDate, taskList) {