5.3.6
focused on .droppable  and how to remove items.

$("#trash").droppable({
adds the trash id to js and assigns a .droppable listener

  accept: ".card .list-group-item",
  tolerance: "touch",
has a acceptance criteria of needing to match those two classes and a tolerance of touch meaning the draggable overlaps the droppable by any amount.

  drop: function(event, ui) {
    ui.draggable.remove();
    console.log("drop");
  },
drop event that uses ui.draggable to target the object being targeted and then assigns it the remove function

  over: function(event, ui) {
    console.log("over");
  },
  out: function(event, ui) {
    console.log("out");
  }
});